--- glym-app/src/features/auth/AuthForm.jsx
@@ -1,151 +1,0 @@
-import { useState, useEffect } from "react";
-import { font } from "../../styles/font";
-import logo from "../../shared/GLYM_LOGO.png"
-import kakaoLogo from "../../shared/KAKAO_LOGO.png"
-import styled from "styled-components";
-import { lighten, darken } from 'polished';
-
-const Container = styled.div`
-    width: 584px;
-    height: 700px;
-
-    margin: 0 auto;
-    text-align: center;
-
-    border: 1px solid #D9D9D9;
-    border-radius: 10px;
-`;
-
-const Input = styled.input`
-    width: 400px;
-    height: 50px;
-    box-sizing: border-box;
-
-    display: block;
-    padding: 0 30px;
-    margin: 0 auto;
-    margin-bottom: 15px;
-
-    ${font(16, 400, 1.5)}
-
-    border: 1px solid #929292;
-    border-radius: 10px;
-
-    &:focus {
-        outline: none;
-        border: 1px solid #FF3F77;
-    }
-`;
-
-const LoginButton = styled.button`
-    width: 400px;
-    height: 60px;
-
-    margin-top: ${(props) => props.$marginTop || '15px;'};
-
-    ${font(20, 400, 1.5)}
-    color: ${(props) => props.$fontColor || '#000000'};
-    background-color: ${(props) => props.$bgColor || '#FFFFFF'};
-
-    border: none;
-    border-radius: 10px;
-
-    &:hover {
-        background-color: ${(props) => darken(0.05, props.$bgColor || '#FFFFFF')};
-    }
-`;
-
-const AuthUtilityButtonContainer = styled.div`
-    width: 400px; 
-
-    margin: 0 auto;
-    margin-top: 10px;
-
-    display: flex;
-    justify-content: space-between;
-
-    & > :nth-child(1) { margin-right: auto; }
-    & > :nth-child(2), & > :nth-child(3) { margin-left: 15px; }
-`;
-
-const AuthUtilityButton = styled.button`
-    background-color: inherit;
-    color: #6B6B6B;
-
-    border: none;
-
-    &:hover {
-        color: ${lighten(0.2, '#6B6B6B')};
-    }
-`;
-
-const KakaoLogo = styled.img`
-    width: 30px;
-
-    position: absolute;
-    left: 16px;
-    top: 50%; 
-    transform: translateY(-50%);
-`;
-
-const S = {
-    Container, Input, LoginButton, AuthUtilityButton, AuthUtilityButtonContainer, KakaoLogo
-};
-
-export default function AuthForm() {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
-
-    function submitLogin(username, password) {
-        console.log("로그인 버튼 클릭 확인");
-    }
-
-    return (
-        <>
-            <S.Container>
-                <img src={logo} alt="glym_logo" style={{ width: "150px", margin: "30px 0 60px 0" }}></img>
-
-                <form onSubmit={(e) => {
-                    e.preventDefault();
-                    submitLogin(username, password);
-                }} >
-                    <S.Input
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                        type="email"
-                        placeholder="이메일"
-                        required
-                    />
-                    <S.Input
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                        type="password"
-                        placeholder="비밀번호"
-                        required
-                    />
-
-                    <S.LoginButton
-                        type="submit"
-                        $bgColor="#FF3F77"
-                        $fontColor="#FFFFFF"
-                    >
-                        로그인
-                    </S.LoginButton>
-                </form>
-
-                <S.AuthUtilityButtonContainer>
-                    <S.AuthUtilityButton>회원가입</S.AuthUtilityButton>
-                    <S.AuthUtilityButton>아이디 찾기</S.AuthUtilityButton>
-                    <S.AuthUtilityButton>비밀번호 찾기</S.AuthUtilityButton>
-                </S.AuthUtilityButtonContainer>
-
-                <hr style={{ width: "400px", margin: "0 auto", marginTop: "5px" }} />
-
-                <S.LoginButton $bgColor="#FEE500" $fontColor="#000000" $marginTop="30px" style={{ position: "relative" }}>
-                    <S.KakaoLogo src={kakaoLogo} />
-                    카카오 로그인
-                </S.LoginButton>
-            </S.Container>
-        </>
-    );
-}
