+++ glym-app/src/features/auth/components/AuthForm.jsx
@@ -1,0 +1,62 @@
+import { useState, useEffect } from "react";
+import { S } from '../style';
+import logo from "../../../shared/GLYM_LOGO.png"
+import kakaoLogo from "../../../shared/KAKAO_LOGO.png"
+
+export default function AuthForm() {
+    const [username, setUsername] = useState('');
+    const [password, setPassword] = useState('');
+
+    function submitLogin(username, password) {
+        console.log("로그인 버튼 클릭 확인");
+    }
+
+    return (
+        <>
+            <S.Auth.Container>
+                <img src={logo} alt="glym_logo" style={{ width: "150px", margin: "30px 0 60px 0" }}></img>
+
+                <form onSubmit={(e) => {
+                    e.preventDefault();
+                    submitLogin(username, password);
+                }} >
+                    <S.Auth.Input
+                        value={username}
+                        onChange={(e) => setUsername(e.target.value)}
+                        type="email"
+                        placeholder="이메일"
+                        required
+                    />
+                    <S.Auth.Input
+                        value={password}
+                        onChange={(e) => setPassword(e.target.value)}
+                        type="password"
+                        placeholder="비밀번호"
+                        required
+                    />
+
+                    <S.Auth.LoginButton
+                        type="submit"
+                        $bgColor="#FF3F77"
+                        $fontColor="#FFFFFF"
+                    >
+                        로그인
+                    </S.Auth.LoginButton>
+                </form>
+
+                <S.Auth.AuthUtilityButtonContainer>
+                    <S.Auth.AuthUtilityButton>회원가입</S.Auth.AuthUtilityButton>
+                    <S.Auth.AuthUtilityButton>아이디 찾기</S.Auth.AuthUtilityButton>
+                    <S.Auth.AuthUtilityButton>비밀번호 찾기</S.Auth.AuthUtilityButton>
+                </S.Auth.AuthUtilityButtonContainer>
+
+                <hr style={{ width: "400px", margin: "0 auto", marginTop: "5px" }} />
+
+                <S.Auth.LoginButton $bgColor="#FEE500" $fontColor="#000000" $marginTop="30px" style={{ position: "relative" }}>
+                    <S.Auth.KakaoLogo src={kakaoLogo} />
+                    카카오 로그인
+                </S.Auth.LoginButton>
+            </S.Auth.Container>
+        </>
+    );
+}
